requirements_gcp.txt
transformers
pandas
datasets
faker
nlpaug
torch
google-cloud-storage
tensorboard
accelerate 